@@ -5,7 +5,7 @@ homepage: https://github.com/fluttercommunity/firebase_dart_sdk/tree/develop/goo
 commit: 89794a635a5efad6705e3d7ccc29e8139318d947
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
