@@ -10,18 +10,14 @@ part of '../google_sign_in_dartio.dart';
 /// Using this implementation will not provide a refresh token, that means that
 /// the user will need to login again after the access token expires (~1 hour)
 Future<Map<String, dynamic>> _tokenSignIn({
-  @required String clientId,
-  @required String scope,
-  @required UrlPresenter presenter,
-  String hostedDomains,
-  String uid,
-  String successUrl,
-  String failUrl,
+  required String clientId,
+  required String scope,
+  required UrlPresenter presenter,
+  String? hostedDomains,
+  String? uid,
+  String? successUrl,
+  String? failUrl,
 }) async {
-  assert(presenter != null);
-  assert(clientId != null);
-  assert(scope != null);
-
   final Completer<Map<String, dynamic>> completer =
       Completer<Map<String, dynamic>>();
 
@@ -40,7 +36,7 @@ Future<Map<String, dynamic>> _tokenSignIn({
     if (uri.path == '/') {
       return _sendData(request, _verifyFragmentHtml);
     } else if (uri.path == '/response') {
-      if (successUrl.isNotEmpty && failUrl.isNotEmpty) {
+      if (successUrl!.isNotEmpty && failUrl!.isNotEmpty) {
         await _validateTokenWithCustomScreen(
                 request, state, successUrl, failUrl)
             .then(completer.complete)
@@ -86,10 +82,10 @@ Future<Map<String, dynamic>> _tokenSignIn({
 Future<Map<String, String>> _validateTokenResponse(
     HttpRequest request, String state) async {
   final Map<String, String> authResponse = request.requestedUri.queryParameters;
-  final String returnedState = authResponse['state'];
-  final String accessToken = authResponse['access_token'];
-  final String idToken = authResponse['id_token'];
-  String message;
+  final String? returnedState = authResponse['state'];
+  final String? accessToken = authResponse['access_token'];
+  final String? idToken = authResponse['id_token'];
+  String? message;
   if (state != returnedState) {
     message = 'Invalid response from server (state did not match).';
   }
@@ -109,11 +105,11 @@ Future<Map<String, String>> _validateTokenResponse(
 }
 
 Future<Map<String, String>> _validateTokenWithCustomScreen(HttpRequest request,
-    String state, String successUrl, String failUrl) async {
+    String state, String? successUrl, String? failUrl) async {
   final Map<String, String> authResponse = request.requestedUri.queryParameters;
-  final String returnedState = authResponse['state'];
-  final String accessToken = authResponse['access_token'];
-  final String idToken = authResponse['id_token'];
+  final String? returnedState = authResponse['state'];
+  final String? accessToken = authResponse['access_token'];
+  final String? idToken = authResponse['id_token'];
   if (state != returnedState ||
       accessToken == null ||
       accessToken.isEmpty ||
@@ -123,12 +119,12 @@ Future<Map<String, String>> _validateTokenWithCustomScreen(HttpRequest request,
       ..statusCode = 500
       ..headers.set('content-type', 'text/plain')
       ..write('');
-    await launch(failUrl);
+    await launch(failUrl!);
   } else {
     request.response
       ..statusCode = 200
       ..write('');
-    await launch(successUrl);
+    await launch(successUrl!);
   }
   await request.response.close();
   return authResponse;
